requirements.txt
celery==5.3.1
docker==6.1.3
channels==4.0.0
django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-json-widget==1.1.1
drf-spectacular==0.26.4
gunicorn==21.2.0
nginxfmt==1.2.2
paramiko==3.3.1
PyMySQL==1.1.0
python-decouple==3.8
requests==2.31.0
Faker==19.3.0
uvicorn[standard]
paho-mqtt==1.6.1
python-dotenv==1.0.0